+++ kore/src/Kore/Repl.hs
@@ -1,12 +1,10 @@
 {-|
 Module      : Kore.Repl
-Description : Logging functions.
+Description : REPL
 Copyright   : (c) Runtime Verification, 2019
 License     : NCSA
 Maintainer  : vladimir.ciobanu@runtimeverification.com
 -}
-
-{-# LANGUAGE TemplateHaskell #-}
 
 module Kore.Repl
     ( runRepl
@@ -19,7 +17,6 @@ import           Control.Lens
                  ( (.=) )
 import qualified Control.Lens as Lens hiding
                  ( makeLenses )
-import qualified Control.Lens.TH.Rules as Lens
 import           Control.Monad.Catch
                  ( MonadCatch, catch )
 import           Control.Monad.Extra
@@ -32,54 +29,29 @@ import           Control.Monad.State.Strict
                  ( evalStateT, get )
 import           Control.Monad.State.Strict
                  ( lift )
-import           Data.Functor
-                 ( ($>) )
-import           Data.Graph.Inductive.Graph
-                 ( Graph )
 import qualified Data.Graph.Inductive.Graph as Graph
-import qualified Data.GraphViz as Graph
 import           Data.Maybe
                  ( listToMaybe )
-import           Data.Text.Prettyprint.Doc
-                 ( pretty )
 import           System.IO
                  ( hFlush, stdout )
 import           Text.Megaparsec
-                 ( Parsec, option, optional, parseMaybe, (<|>) )
-import           Text.Megaparsec.Char
-import           Text.Megaparsec.Char.Lexer
-                 ( decimal, signed )
+                 ( parseMaybe )
 
-import           Control.Monad.Extra
-                 ( loopM )
-import qualified Kore.AST.Common as Kore
-                 ( Variable )
 import           Kore.AST.MetaOrObject
                  ( MetaOrObject )
-import           Kore.Attribute.Axiom
-                 ( SourceLocation (..) )
-import qualified Kore.Attribute.Axiom as Attribute
-                 ( sourceLocation )
 import           Kore.IndexedModule.MetadataTools
                  ( MetadataTools )
-import           Kore.OnePath.Step
-                 ( CommonStrategyPattern, StrategyPattern (..) )
 import           Kore.OnePath.Verification
                  ( verifyClaimStep )
 import           Kore.OnePath.Verification
-                 ( Axiom (..) )
+                 ( Axiom )
 import           Kore.OnePath.Verification
-                 ( Claim (..) )
+                 ( Claim )
+import           Kore.Repl.Data
+import           Kore.Repl.Interpreter
+import           Kore.Repl.Parser
 import           Kore.Step.Axiom.Data
                  ( BuiltinAndAxiomSimplifierMap )
-import           Kore.Step.AxiomPatterns
-                 ( RewriteRule (..) )
-import           Kore.Step.AxiomPatterns
-                 ( RulePattern (..) )
-import qualified Kore.Step.AxiomPatterns as Axiom
-                 ( attributes )
-import           Kore.Step.Representation.ExpandedPattern
-                 ( Predicated (..) )
 import           Kore.Step.Simplification.Data
                  ( Simplifier )
 import           Kore.Step.Simplification.Data
@@ -89,29 +61,6 @@ import           Kore.Step.Simplification.Data
 import           Kore.Step.StepperAttributes
                  ( StepperAttributes )
 import qualified Kore.Step.Strategy as Strategy
-import           Kore.Unparser
-                 ( unparseToString )
-
--- Type synonym for the actual type of the execution graph.
-type ExecutionGraph level = Strategy.ExecutionGraph (CommonStrategyPattern level)
-
--- | State for the rep.
-data ReplState level = ReplState
-    { axioms  :: [Axiom level]
-    -- ^ List of available axioms
-    , claims  :: [Claim level]
-    -- ^ List of claims to be proven
-    , claim   :: Claim level
-    -- ^ Currently focused claim in the repl
-    , graph   :: ExecutionGraph level
-    -- ^ Execution graph for the current proof; initialized with root = claim
-    , node    :: Graph.Node
-    -- ^ Currently selected node in the graph; initialized with node = root
-    , stepper :: StateT (ReplState level) Simplifier Bool
-    -- ^ Stepper function, it is a partially applied 'verifyClaimStep'
-    }
-
-Lens.makeLenses ''ReplState
 
 -- | Runs the repl for proof mode. It requires all the tooling and simplifiers
 -- that would otherwise be required in the proof and allows for step-by-step
@@ -200,290 +149,3 @@ runRepl tools simplifier predicateSimplifier axiomToIdSimplifier axioms' claims'
             putStr $ "Kore (" <> show node <> ")> "
             hFlush stdout
             getLine
-
--- | List of available commands for the Repl. Note that we are always in a proof
--- state. We pick the first available Claim when we initialize the state.
-data ReplCommand
-    = ShowUsage
-    -- ^ This is the default action in case parsing all others fail.
-    | Help
-    -- ^ Shows the help message.
-    | ShowClaim !Int
-    -- ^ Show the nth claim.
-    | ShowAxiom !Int
-    -- ^ Show the nth axiom.
-    | Prove !Int
-    -- ^ Drop the current proof state and re-initialize for the nth claim.
-    | ShowGraph
-    -- ^ Show the current execution graph.
-    | ProveSteps !Int
-    -- ^ Do n proof steps from curent node.
-    | SelectNode !Int
-    -- ^ Select a different node in the graph.
-    | ShowConfig !(Maybe Int)
-    -- ^ Show the configuration from the current node.
-    | Exit
-    -- ^ Exit the repl.
-
--- | Please remember to update this text whenever you update the ADT above.
-helpText :: String
-helpText =
-    "Available commands in the Kore REPL: \n\
-    \help                    shows this help message\n\
-    \claim <n>               shows the nth claim\n\
-    \axiom <n>               shows the nth axiom\n\
-    \prove <n>               initializez proof mode for the nth \
-                             \claim\n\
-    \graph                   shows the current proof graph\n\
-    \step [n]                attempts to run 'n' proof steps at\
-                             \the current node (n=1 by default)\n\
-    \select <n>              select node id 'n' from the graph\n\
-    \config [n]              shows the config for node 'n'\
-                             \(defaults to current node)\n\
-    \exit                    exits the repl"
-
-type Parser = Parsec String String
-
-commandParser :: Parser ReplCommand
-commandParser =
-    help0
-    <|> showClaim0
-    <|> showAxiom0
-    <|> prove0
-    <|> showGraph0
-    <|> proveSteps0
-    <|> selectNode0
-    <|> showConfig0
-    <|> exit0
-    <|> pure ShowUsage
-  where
-    help0 :: Parser ReplCommand
-    help0 = string "help" $> Help
-
-    showClaim0 :: Parser ReplCommand
-    showClaim0 = fmap ShowClaim $ string "claim" *> space *> decimal
-
-    showAxiom0 :: Parser ReplCommand
-    showAxiom0 = fmap ShowAxiom $ string "axiom" *> space *> decimal
-
-    prove0 :: Parser ReplCommand
-    prove0 = fmap Prove $ string "prove" *> space *> decimal
-
-    showGraph0 :: Parser ReplCommand
-    showGraph0 = ShowGraph <$ string "graph"
-
-    proveSteps0 :: Parser ReplCommand
-    proveSteps0 = fmap ProveSteps $ string "step" *> space *> option 1 decimal
-
-    selectNode0 :: Parser ReplCommand
-    selectNode0 =
-        fmap SelectNode $ string "select" *> space *> signed space decimal
-
-    showConfig0 :: Parser ReplCommand
-    showConfig0 = fmap ShowConfig $ string "config" *> optional (space *> decimal)
-
-    exit0 :: Parser ReplCommand
-    exit0 = Exit <$ string "exit"
-
-replInterpreter
-    :: forall level
-    .  MetaOrObject level
-    => ReplCommand
-    -> StateT (ReplState level) Simplifier Bool
-replInterpreter =
-    \case
-        ShowUsage -> showUsage0 $> True
-        Help -> help0 $> True
-        ShowClaim c -> showClaim0 c $> True
-        ShowAxiom a -> showAxiom0 a $> True
-        Prove i -> prove0 i $> True
-        ShowGraph -> showGraph0 $> True
-        ProveSteps n -> proveSteps0 n $> True
-        SelectNode i -> selectNode0 i $> True
-        ShowConfig mc -> showConfig0 mc $> True
-        Exit -> pure False
-  where
-    showUsage0 :: MonadIO m => m ()
-    showUsage0 =
-        putStrLn' "Could not parse command, try using 'help'."
-
-    help0 :: MonadIO m => m ()
-    help0 = putStrLn' helpText
-
-    showClaim0
-        :: MonadIO m
-        => MonadState (ReplState level) m
-        => Int
-        -> m ()
-    showClaim0 index = do
-        claim <- Lens.preuse $ lensClaims . Lens.element index
-        maybe printNotFound (printRewriteRule . unClaim) $ claim
-
-    showAxiom0
-        :: MonadIO m
-        => MonadState (ReplState level) m
-        => Int
-        -> m ()
-    showAxiom0 index = do
-        axiom <- Lens.preuse $ lensAxioms . Lens.element index
-        maybe printNotFound (printRewriteRule . unAxiom) $ axiom
-
-    prove0
-        :: MonadIO m
-        => MonadState (ReplState level) m
-        => Int
-        -> m ()
-    prove0 index = do
-        claim' <- Lens.preuse $ lensClaims . Lens.element index
-        case claim' of
-            Nothing -> printNotFound
-            Just claim -> do
-                let
-                    graph@Strategy.ExecutionGraph { root }
-                        = emptyExecutionGraph claim
-                lensGraph .= graph
-                lensClaim .= claim
-                lensNode  .= root
-                putStrLn' "Execution Graph initiated"
-
-    showGraph0
-        :: MonadIO m
-        => MonadState (ReplState level) m
-        => m ()
-    showGraph0 = do
-        Strategy.ExecutionGraph { graph } <- Lens.use lensGraph
-        liftIO $ showDotGraph graph
-
-    proveSteps0 :: Int -> StateT (ReplState level) Simplifier ()
-    proveSteps0 n = do
-        result <- loopM performStepNoBranching (n, Success)
-        case result of
-            (0, Success) -> pure ()
-            (done, res) ->
-                putStrLn'
-                    $ "Stopped after "
-                    <> show (n - done - 1)
-                    <> " step(s) due to "
-                    <> show res
-
-    selectNode0 :: Int -> StateT (ReplState level) Simplifier ()
-    selectNode0 i = do
-        Strategy.ExecutionGraph { graph } <- Lens.use lensGraph
-        if i `elem` Graph.nodes graph
-            then lensNode .= i
-            else putStrLn' "Invalid node!"
-
-    showConfig0 :: Maybe Int -> StateT (ReplState level) Simplifier ()
-    showConfig0 configNode = do
-        Strategy.ExecutionGraph { graph } <- Lens.use lensGraph
-        node <- Lens.use lensNode
-        let node' = maybe node id configNode
-        if node' `elem` Graph.nodes graph
-           then do
-               putStrLn' $ "Config at node " <> show node' <> " is:"
-               putStrLn'
-                   . unparseStrategy
-                   . Graph.lab'
-                   . Graph.context graph
-                   $ node'
-           else putStrLn' "Invalid node!"
-
-    printRewriteRule :: MonadIO m => RewriteRule level Kore.Variable -> m ()
-    printRewriteRule rule = do
-        putStrLn' $ unparseToString rule
-        putStrLn'
-            . show
-            . pretty
-            . extractSourceAndLocation
-            $ rule
-
-    performSingleStep
-        :: StateT (ReplState level) Simplifier StepResult
-    performSingleStep = do
-        f <- Lens.use lensStepper
-        node <- Lens.use lensNode
-        res <- f
-        if res
-            then do
-                Strategy.ExecutionGraph { graph } <- Lens.use lensGraph
-                let
-                    context = Graph.context graph node
-                case Graph.suc' context of
-                    [] -> pure NoChildNodes
-                    [configNo] -> do
-                        lensNode .= configNo
-                        pure Success
-                    neighbors -> pure (Branch neighbors)
-            else pure NodeAlreadyEvaluated
-
-    -- | Performs n proof steps, picking the next node unless branching occurs.
-    -- Returns 'Left' while it has to continue looping, and 'Right' when done
-    -- or when execution branches or proof finishes earlier than the counter.
-    --
-    -- See 'loopM' for details.
-    performStepNoBranching
-        :: (Int, StepResult)
-        -- ^ (current step, last result)
-        -> StateT
-            (ReplState level)
-            Simplifier
-                (Either
-                     (Int, StepResult)
-                     (Int, StepResult)
-                )
-    performStepNoBranching (0, res) =
-        pure $ Right (0, res)
-    performStepNoBranching (n, Success) = do
-        res <- performSingleStep
-        pure $ Left (n-1, res)
-    performStepNoBranching (n, res) =
-        pure $ Right (n, res)
-
-    unparseStrategy :: CommonStrategyPattern level -> String
-    unparseStrategy =
-        \case
-            Bottom -> "Reached bottom"
-            Stuck pat -> "Stuck: \n" <> unparseToString pat
-            RewritePattern pat -> unparseToString pat
-
-    extractSourceAndLocation
-        :: RewriteRule level Kore.Variable
-        -> SourceLocation
-    extractSourceAndLocation
-        (RewriteRule (RulePattern{ Axiom.attributes })) =
-            Attribute.sourceLocation attributes
-
-    printNotFound :: MonadIO m => m ()
-    printNotFound = putStrLn' "Variable or index not found"
-
-    putStrLn' :: MonadIO m => String -> m ()
-    putStrLn' = liftIO . putStrLn
-
-    showDotGraph :: Graph gr => gr nl el -> IO ()
-    showDotGraph =
-        (flip Graph.runGraphvizCanvas') Graph.Xlib
-            . Graph.graphToDot Graph.nonClusteredParams
-
-data StepResult
-    = NodeAlreadyEvaluated
-    | NoChildNodes
-    | Branch [Graph.Node]
-    | Success
-    deriving Show
-
-unClaim :: forall level. Claim level -> RewriteRule level Kore.Variable
-unClaim Claim { rule } = rule
-
-unAxiom :: Axiom level -> RewriteRule level Kore.Variable
-unAxiom (Axiom rule) = rule
-
-emptyExecutionGraph
-    :: Claim level
-    -> Strategy.ExecutionGraph (CommonStrategyPattern level)
-emptyExecutionGraph = Strategy.emptyExecutionGraph . extractConfig . unClaim
-
-extractConfig
-    :: RewriteRule level Kore.Variable
-    -> CommonStrategyPattern level
-extractConfig (RewriteRule RulePattern { left, requires }) =
-    RewritePattern $ Predicated left requires mempty
